Add support for OAuth authentication.
Basically when an OAuth Authorization header is detected we pass
that header along to the FreshBooks server without doing much
else.

In order to determine where we're sending the request, this patch
also makes the wrapper look in the Authorization header for
a "freshbooks_domain" parameter (in my case using Scribe, the
parameter is "oauth_freshbooks_domain" but the regex allows
for both). I figured this was the cleanest solution, as it
beats muddying up the path or querystring with extra
information.

Signed-off-by: Eddie Ringle <eddie@eringle.net>